Move up severity
