+++ src/main/kotlin/ffc/entity/healthcare/Severity.kt
@@ -1,0 +1,10 @@
+package ffc.entity.healthcare
+
+enum class Severity {
+    UNDEFINED,
+    OK,
+    LOW,
+    MID,
+    HI,
+    VERY_HI,
+}
