+++ src/main/kotlin/ffc/entity/healthcare/analyze/ADLAnalyzer.kt
@@ -18,6 +18,7 @@
 package ffc.entity.healthcare.analyze
 
 import ffc.entity.healthcare.Service
+import ffc.entity.healthcare.Severity
 import ffc.entity.healthcare.SpecialPP
 
 class ADLAnalyzer : Analyzer {
@@ -27,9 +28,9 @@ class ADLAnalyzer : Analyzer {
     override fun analyzeFrom(service: Service): HealthIssue? {
         return if (service is SpecialPP) {
             when (service.ppType.id) {
-                "1B1280" -> HealthProblem(forIssue, service, HealthIssue.Severity.OK)
-                "1B1281" -> HealthProblem(forIssue, service, HealthIssue.Severity.MID)
-                "1B1282" -> HealthProblem(forIssue, service, HealthIssue.Severity.VERY_HI)
+                "1B1280" -> HealthProblem(forIssue, service, Severity.OK)
+                "1B1281" -> HealthProblem(forIssue, service, Severity.MID)
+                "1B1282" -> HealthProblem(forIssue, service, Severity.VERY_HI)
                 else -> null
             }
         } else null
