+++ src/main/kotlin/ffc/entity/healthcare/analyze/CvdAnalyzer.kt
@@ -18,6 +18,7 @@
 package ffc.entity.healthcare.analyze
 
 import ffc.entity.healthcare.Service
+import ffc.entity.healthcare.Severity
 import ffc.entity.healthcare.SpecialPP
 
 class CvdAnalyzer : Analyzer {
@@ -28,10 +29,10 @@ class CvdAnalyzer : Analyzer {
         return when (service) {
             is SpecialPP -> {
                 when (service.ppType.id) {
-                    "1B1230" -> HealthProblem(forIssue, service, HealthIssue.Severity.OK)
-                    "1B1231" -> HealthProblem(forIssue, service, HealthIssue.Severity.MID)
-                    "1B1232" -> HealthProblem(forIssue, service, HealthIssue.Severity.HI)
-                    "1B1234", "1B1235" -> HealthProblem(forIssue, service, HealthIssue.Severity.VERY_HI)
+                    "1B1230" -> HealthProblem(forIssue, service, Severity.OK)
+                    "1B1231" -> HealthProblem(forIssue, service, Severity.MID)
+                    "1B1232" -> HealthProblem(forIssue, service, Severity.HI)
+                    "1B1234", "1B1235" -> HealthProblem(forIssue, service, Severity.VERY_HI)
                     else -> null
                 }
             }
