+++ src/main/kotlin/ffc/entity/healthcare/analyze/DepressiveAnalyzer.kt
@@ -18,6 +18,7 @@
 package ffc.entity.healthcare.analyze
 
 import ffc.entity.healthcare.Service
+import ffc.entity.healthcare.Severity
 import ffc.entity.healthcare.SpecialPP
 
 class DepressiveAnalyzer : Analyzer {
@@ -27,11 +28,11 @@ class DepressiveAnalyzer : Analyzer {
     override fun analyzeFrom(service: Service): HealthIssue? {
         return if (service is SpecialPP) {
             when (service.ppType.id) {
-                "1B0282", "1B0280" -> HealthProblem(forIssue, service, HealthIssue.Severity.OK)
-                "1B0281" -> HealthProblem(forIssue, service, HealthIssue.Severity.UNDEFINED)
-                "1B0283" -> HealthProblem(forIssue, service, HealthIssue.Severity.LOW)
-                "1B0284" -> HealthProblem(forIssue, service, HealthIssue.Severity.MID)
-                "1B0285" -> HealthProblem(forIssue, service, HealthIssue.Severity.VERY_HI)
+                "1B0282", "1B0280" -> HealthProblem(forIssue, service, Severity.OK)
+                "1B0281" -> HealthProblem(forIssue, service, Severity.UNDEFINED)
+                "1B0283" -> HealthProblem(forIssue, service, Severity.LOW)
+                "1B0284" -> HealthProblem(forIssue, service, Severity.MID)
+                "1B0285" -> HealthProblem(forIssue, service, Severity.VERY_HI)
                 else -> null
             }
         } else null
